Extract redaction to utils and redact JSON by walking it
Widening redaction to a tool call's arguments put the inline secret pattern in
front of a shape it had never been shown: a JSON string nested inside another.
Six failures came out of that, one at a time, and each patch bought one shape
and left the next — a nested value's contents, every line of a multi-line .env
but the first, a redacted line swallowing the lines below it, a serialized
backslash read as a boundary, a nested JSON object whose key no longer sits
beside its separator, a nested value holding an escaped quote.

They are one cause, and it is level rather than coverage. The payload that
leaks when serialized is redacted correctly when raw. A regex cannot count
nesting depth.

So redact structure as structure. `arguments` and `result` arrive as JsonValue;
`redactJson` walks them before serialization, destroying any value whose key
names a secret and passing every string leaf through the text rules. Whatever
parsed the structure has already peeled one level of escaping, so the text the
pattern sees is always ordinary text. Depth stops mattering, a payload that
arrived as JSON is still JSON afterwards, and a low-entropy secret under a
secret-named key is covered for the first time. The pattern itself needed no
change and is next's, untouched.

The capability moves to @excitedjs/dreamux-utils, where it is pure and core no
longer owns rules it only consumes. Its secret key names had been written out
three times and the copies had drifted: the host logger's list was missing
api_key, private_key, and client_secret, so those three were never hidden in a
log. One list closes that gap with it.
