+++ common/changes/@excitedjs/dreamux-utils/extract-redaction-capability_2026-09-11-09-50.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "comment": "Add the shared redaction capability: one list of secret key names, the text rules (secret shapes destroyed, host paths renamed), and redactJson, which redacts a JSON value by walking its structure instead of pattern-matching its serialization.",
+      "type": "minor",
+      "packageName": "@excitedjs/dreamux-utils"
+    }
+  ],
+  "packageName": "@excitedjs/dreamux-utils",
+  "email": "hvjfccdy6s@privaterelay.appleid.com"
+}
