+++ common/changes/@excitedjs/dreamux/redact-tool-call-arguments_2026-09-11-07-55.json
@@ -1,7 +1,7 @@
 {
   "changes": [
     {
-      "comment": "Redact tool call invocations and arguments in the conversation projection. Secret masking and workspace/home path renaming now apply to every displayed member of a tool call, and the secret pattern recognizes secrets written inside a nested JSON string, so a token pasted into a command or a .env written through a tool argument no longer reaches a chat surface unmasked.",
+      "comment": "Redact tool call invocations and arguments in the conversation projection. Secret masking and workspace/home path renaming now apply to every displayed member of a tool call, and a call's structured payloads are redacted by walking the JSON value, so a secret written into a nested object or a multi-line .env is masked and the payload is still valid JSON. The redaction rules themselves moved to @excitedjs/dreamux-utils, where the host logger and config display now share one list of secret key names; the logger previously missed api_key, private_key, and client_secret.",
       "type": "minor",
       "packageName": "@excitedjs/dreamux"
     }
