+++ packages/dreamux/src/channel/conversation-projection.ts
@@ -9,66 +9,12 @@ import type {
   TeammateRole,
 } from '@excitedjs/dreamux-types';
 
+import { redactJson, redactText } from '@excitedjs/dreamux-utils';
+
 import { errorInfo } from '../platform/error-info.js';
 import type { DispatcherCoreEventPublisher } from '../service/dispatcher-core-events/index.js';
 import type { AgentEntityIdentity } from '../service/agent-entity/types.js';
 
-/**
- * `key: value` / `key=value` pairs whose key names a secret. The value is one
- * quoted string or one bare word. Three things count as quotes, because a
- * secret arrives inside all three: a plain JSON string, a shell-style single-
- * or back-quoted one, and a JSON string nested inside another — `\"value\"`,
- * which is what `.env` text or `export FOO="bar"` looks like once it is a
- * tool argument. The nested form has to be matched as a unit; a bare word
- * would stop on its opening backslash and leave the value itself uncovered.
- *
- * A bare word stops at whitespace, a separator, a quote, or a closing bracket,
- * and at a JSON escape — `\"`, `\n`, `\r`, `\t`. Those four are the text
- * that *ends* a value rather than text inside one: the escape closing the
- * string around it, or the line break after it. A backslash that is itself
- * escaped, `\\`, is the opposite — it is one character *of* the value — so it
- * is taken as a unit before those four are looked for, and a serialized
- * `C:\\name` is covered whole rather than cut at the `\n` that is not there.
- *
- * The same four are a boundary at the key end too, where `\b` cannot see one:
- * a newline serialized into JSON is the two characters `\n`, and its `n` is a
- * word character, which would hide every line of a multi-line `.env` but the
- * first. A key still needs a boundary, so `mytoken` is not one.
- *
- * Together these keep the shape *around* the secret intact, so a payload that
- * arrived as JSON is still JSON afterwards and the Channel can parse it to
- * decide how to show it, and a redacted line does not take the lines after it
- * along.
- */
-const INLINE_SECRET_RE = /(["']?(?:\b|(?<=\\[nrt]))(?:secret|password|passwd|token|authorization|cookie|credential|api[_-]?key|private[_-]?key|client[_-]?secret)\b["']?)(\s*[:=]\s*)(\\"(?:[^"\\]|\\.)*?\\"|"(?:[^"\\]|\\.)*"|'[^']*'|`[^`]*`|(?:\\\\|(?!\\["nrt])[^\s,;"'`)\]}])+)/giu;
-const BEARER_RE = /\bBearer\s+[A-Za-z0-9._~+/-]+=*/giu;
-const PRIVATE_KEY_RE = /-----BEGIN [A-Z0-9 ]*PRIVATE KEY-----[\s\S]*?-----END [A-Z0-9 ]*PRIVATE KEY-----/giu;
-const JWT_RE = /\beyJ[A-Za-z0-9_-]{8,}\.[A-Za-z0-9_-]{8,}\.[A-Za-z0-9_-]{8,}\b/gu;
-const COMMON_ACCESS_KEY_RE = /\b(?:AKIA|ASIA|AKLT)[A-Z0-9]{12,}\b/gu;
-
-/**
- * What may sit inside a path without ending it.
- *
- * A prefix only counts as a path when the characters around it agree that it is
- * one: `~/work` is this operator's home, `/home/alicexyz` is somebody else's
- * directory that merely starts with the same letters, and `not/home/alice` is a
- * fragment of a longer path that was never rooted here. Letters, digits, and
- * the separators/punctuation that appear inside real path segments continue a
- * token; anything else — whitespace, a quote, a colon, a comma — ends it.
- * Two narrower exceptions are handled at a match: a period closes prose only
- * when what follows is already a boundary, and a preceding slash starts a path
- * only when it completes a URL scheme's `://`.
- */
-const PATH_TOKEN_CHARACTER_RE = /[\p{L}\p{N}_.~\\/-]/u;
-
-/**
- * Core redacts and never truncates. How much of a payload a surface can show
- * is that surface's own limit, applied where it sends: the Channel parses the
- * JSON first and cuts just before the send call. Cutting here would hand
- * every surface an already-damaged value — a JSON result that no longer
- * parses — with no way to get it back.
- */
-interface RedactedText { value: string; redacted: boolean }
 
 /**
  * The projected Agent: its durable identity plus the runtime role its owner
@@ -229,13 +175,16 @@ function projectedActivity(
       };
     }
     case 'tool.call': {
-      const redact = (text: string | null): RedactedText | null =>
+      const redact = (text: string | null) =>
         text === null ? null : redactText(text, cwd, homePathPrefixes);
       const summary = redact(activity.summary);
       const invocation = redact(activity.invocation);
-      const args = redact(jsonText(activity.arguments));
-      const result = redact(jsonText(activity.error ?? activity.result));
       const items = activity.items.map((item) => redactText(item, cwd, homePathPrefixes));
+      // A call's payloads arrive as structure, so they are walked rather than
+      // read as one string: a key that names a secret is answered by its name,
+      // and every string leaf is ordinary text by the time it is redacted.
+      const args = redactJson(activity.arguments, cwd, homePathPrefixes);
+      const result = redactJson(activity.error ?? activity.result, cwd, homePathPrefixes);
       return {
         kind: 'tool.call',
         event_id: activity.id,
@@ -246,8 +195,8 @@ function projectedActivity(
         invocation: invocation?.value ?? null,
         items: items.map((item) => item.value),
         status: activity.status,
-        arguments_json: args?.value ?? null,
-        result_json: result?.value ?? null,
+        arguments_json: jsonText(args.value),
+        result_json: jsonText(result.value),
         redacted: [summary, invocation, args, result, ...items]
           .some((member) => member?.redacted ?? false),
       };
@@ -268,120 +217,11 @@ function projectedActivity(
 
 /**
  * A structured value travels as its compact JSON text; a value that already is
- * a string travels as itself. What this guarantees is the serialization: the
- * text is whole, never cut. Parsability is not part of it — the result text
- * still passes the redactor after this point, and a replacement that lands
- * inside a JSON string can leave text no parser accepts — so a consumer that
- * wants the structure back treats a parse failure as "this is text", which is
- * what a display does with any payload it cannot read as JSON.
+ * a string travels as itself. The text is whole, never cut, and it is written
+ * after redaction rather than before, so a structured payload is still valid
+ * JSON when it arrives.
  */
 function jsonText(value: JsonValue | string | null): string | null {
   if (value === null) return null;
   return typeof value === 'string' ? value : JSON.stringify(value);
-}
-
-/**
- * Rewrite what a projected conversation must not publish verbatim.
- *
- * Two different jobs share this function. Secrets are *destroyed* — a token has
- * no legible form worth keeping. Paths are only *renamed*: an operator reading
- * a card still needs to know which file was touched, so the workspace becomes
- * `.` and this host's home becomes `~`, exactly the way the operator's own shell
- * prints them. Order matters: the workspace usually sits under the home, so
- * relativizing it first keeps the shorter, more useful form.
- *
- * `homePathPrefixes` is explicit so this pure projection never depends on
- * process-global resolution state.
- */
-export function redactText(
-  value: string,
-  cwd: string,
-  homePaths: readonly string[],
-): RedactedText {
-  let redacted = replacePathPrefix(value, cwd, '.', '', true);
-  redacted = redacted.replace(PRIVATE_KEY_RE, '<redacted-private-key>');
-  for (const homePath of homePaths) {
-    redacted = replacePathPrefix(redacted, homePath, '~', '~', false);
-  }
-  redacted = redacted.replace(BEARER_RE, 'Bearer <redacted>');
-  redacted = redacted.replace(JWT_RE, '<redacted-jwt>');
-  redacted = redacted.replace(COMMON_ACCESS_KEY_RE, '<redacted-access-key>');
-  redacted = redacted.replace(
-    INLINE_SECRET_RE,
-    (_match, key: string, separator: string, secret: string) => {
-      // A quoted secret stays a quoted (now empty of meaning) string, so the
-      // text around it keeps whatever grammar it had — JSON included. A nested
-      // JSON string is quoted by its escape, not by the bare character.
-      const quote = secret.startsWith('\\"')
-        ? '\\"'
-        : (/^["'`]/u.test(secret) ? secret[0] : '');
-      return `${key}${separator}${quote}<redacted>${quote}`;
-    },
-  );
-  return { value: redacted, redacted: redacted !== value };
-}
-
-/**
- * Replace every occurrence of `rawPrefix` that is actually the head of a path.
- *
- * Scanning for a known prefix is what makes this honest where a regex is not: a
- * pattern like `/home/<name>/...` matches any string of that *shape*, including
- * a directory on some other machine quoted in a log, and blanking those costs
- * legibility for no privacy gain. Only the prefixes this host really uses are
- * offered here, and each hit must still be bounded on both sides — preceded by
- * a non-path character and followed by a separator or the end of a token.
- *
- * A hit with a path continuing after it (`<prefix>/rest`) takes
- * `nestedReplacement`, and `stripNestedSeparator` drops the separator with it so
- * a workspace `<cwd>/rest` turns into `rest`. A hit that ends there takes
- * `exactReplacement`, so a bare workspace becomes `.`.
- */
-function replacePathPrefix(
-  value: string,
-  rawPrefix: string,
-  exactReplacement: string,
-  nestedReplacement: string,
-  stripNestedSeparator: boolean,
-): string {
-  const prefix = rawPrefix.replace(/[\\/]+$/u, '');
-  if (prefix === '') return value;
-
-  let cursor = 0;
-  let searchFrom = 0;
-  let result = '';
-  while (searchFrom < value.length) {
-    const matchAt = value.indexOf(prefix, searchFrom);
-    if (matchAt < 0) break;
-
-    const suffixAt = matchAt + prefix.length;
-    const next = value[suffixAt];
-    const nested = next === '/' || next === '\\';
-    const ends = isPathPrefixEnd(value, suffixAt);
-    if (isPathPrefixBoundary(value, matchAt) && (nested || ends)) {
-      result += value.slice(cursor, matchAt);
-      result += nested ? nestedReplacement : exactReplacement;
-      cursor = suffixAt + (nested && stripNestedSeparator ? 1 : 0);
-      searchFrom = cursor;
-      continue;
-    }
-    searchFrom = suffixAt;
-  }
-  return cursor === 0 ? value : result + value.slice(cursor);
-}
-
-function isPathPrefixBoundary(value: string, matchAt: number): boolean {
-  if (matchAt === 0 || isPathTokenBoundary(value[matchAt - 1])) return true;
-  return matchAt >= 3 && value.slice(matchAt - 3, matchAt) === '://';
-}
-
-function isPathPrefixEnd(value: string, suffixAt: number): boolean {
-  const next = value[suffixAt];
-  if (next === undefined) return true;
-  return next === '.'
-    ? isPathTokenBoundary(value[suffixAt + 1])
-    : isPathTokenBoundary(next);
-}
-
-function isPathTokenBoundary(character: string | undefined): boolean {
-  return character === undefined || !PATH_TOKEN_CHARACTER_RE.test(character);
 }
