+++ packages/dreamux/tests/cot-projection-privacy.test.ts
@@ -25,7 +25,6 @@ import type {
 
 import {
   createConversationProjection,
-  redactText,
   type ProjectedAgent,
 } from '../src/channel/conversation-projection.js';
 import {
@@ -137,18 +136,6 @@ describe('conversation projection: secret redaction', () => {
     expect(tool.kind === 'tool.call' && tool.redacted).toBe(true);
   });
 
-  it('stops a bare secret at the quote and bracket that close its JSON string', () => {
-    const envelope = JSON.stringify({
-      content: [{ type: 'text', text: 'token: xyz' }],
-      structuredContent: { ok: true },
-    });
-    const { value } = redactText(envelope, '/workspace/repo', ['/home/me']);
-    expect(JSON.parse(value)).toEqual({
-      content: [{ type: 'text', text: 'token: <redacted>' }],
-      structuredContent: { ok: true },
-    });
-  });
-
   it('redacts an inline api_key= assignment', () => {
     const { publisher, projection, agent } = harness();
     const fakeSecret = ['sk', 'live', 'abcdef1234567890'].join('_');
@@ -216,130 +203,15 @@ describe('conversation projection: secret redaction', () => {
     expect(content).toBe('file at secrets/env.local was read');
   });
 
-  it('renders the bare workspace itself as a dot', () => {
-    expect(redactText(`ran in ${CWD} today`, CWD, []).value).toBe('ran in . today');
-  });
-
-  it('renames this host home to ~, keeping the rest of the path legible', () => {
-    expect(redactText('key at /home/me/.ssh/id_rsa now', '', ['/home/me']).value)
-      .toBe('key at ~/.ssh/id_rsa now');
-  });
-
-  it('renames a Windows home the same way', () => {
-    const home = 'C:\\Users\\me';
-    expect(redactText(`backup at ${home}\\notes.txt now`, '', [home]).value)
-      .toBe('backup at ~\\notes.txt now');
-  });
-
-  it('renames the bare home with no path after it', () => {
-    expect(redactText('cd /home/me and stop', '', ['/home/me']).value)
-      .toBe('cd ~ and stop');
-  });
-
-  it('renames a home prefix at the head of a file URL', () => {
-    expect(redactText('open file:///home/me/x', '', ['/home/me']).value)
-      .toBe('open file://~/x');
-  });
-
-  it('renames a bare home before ordinary closing punctuation', () => {
-    const value =
-      'paths /home/me. /home/me, /home/me; /home/me: (/home/me) [/home/me] "/home/me"';
-    expect(redactText(value, '', ['/home/me']).value).toBe(
-      'paths ~. ~, ~; ~: (~) [~] "~"',
-    );
-  });
-
-  it('treats a trailing period as prose punctuation for home and workspace paths', () => {
-    const home = '/home/me';
-    const cwd = `${home}/work/repo`;
-    expect(
-      redactText(`see ${home}. edit ${cwd}/a.ts.`, cwd, [home]).value,
-    ).toBe('see ~. edit a.ts.');
-  });
-
-  it('distinguishes dot-suffixed home siblings from workspace-adjacent siblings', () => {
-    const home = '/home/me';
-    const cwd = `${home}/work/repo`;
-    const value = [
-      `${home}.bak/notes.md`,
-      `${home}.git/config`,
-      `${cwd}.bak/notes.md`,
-      `${cwd}.git/config`,
-    ].join(' ');
-    expect(redactText(value, cwd, [home]).value).toBe([
-      `${home}.bak/notes.md`,
-      `${home}.git/config`,
-      '~/work/repo.bak/notes.md',
-      '~/work/repo.git/config',
-    ].join(' '));
-  });
-
-  it('renames workspace-adjacent siblings through the containing home prefix', () => {
-    const home = '/home/me';
-    const cwd = `${home}/work/repo`;
-    expect(
-      redactText(`${cwd}.git/config ${cwd}-old/x`, cwd, [home]).value,
-    ).toBe('~/work/repo.git/config ~/work/repo-old/x');
-  });
-
-  it('does not treat a doubled filesystem separator as a URL scheme boundary', () => {
-    const home = '/home/me';
-    const cwd = `${home}/work/repo`;
-    const value = `/mnt/backup/${home}/x /mnt/backup/${cwd}/x`;
-    expect(redactText(value, cwd, [home]).value).toBe(value);
-  });
-
   /**
    * The whole reason this is prefix scanning rather than a `/home/<name>/…`
    * regex: another account's directory is not this operator's home, and blanking
    * it costs the reader the one fact they needed.
    */
-  it('leaves a path that merely starts with the same characters alone', () => {
-    const value =
-      'compare /home/mexyz and /home/me-old/notes.txt with /home/meredith/notes.txt';
-    expect(redactText(value, '', ['/home/me']).value).toBe(value);
-  });
-
-  it('leaves a home-shaped fragment that is not rooted alone', () => {
-    const value = 'the string not/home/me/x is not a path here';
-    expect(redactText(value, '', ['/home/me']).value).toBe(value);
-  });
-
-  it('does not treat a foreign home-shaped path as this host home', () => {
-    const value = 'their build ran in /home/someoneelse/repo';
-    expect(redactText(value, '', ['/home/me']).value).toBe(value);
-  });
-
-  it('prefers the longest matching home prefix', () => {
-    expect(
-      redactText('at /home/me/nested/file.ts', '', ['/home/me/nested', '/home/me']).value,
-    ).toBe('at ~/file.ts');
-  });
-
   /**
    * A workspace normally sits under the home. Relativizing it first is what
    * keeps the shorter, more useful form instead of `~/...`-prefixing everything.
    */
-  it('relativizes the workspace before renaming the home it sits under', () => {
-    const home = '/home/me';
-    const cwd = `${home}/work/repo`;
-    expect(
-      redactText(`edited ${cwd}/src/a.ts and ${home}/.config/x`, cwd, [home]).value,
-    ).toBe('edited src/a.ts and ~/.config/x');
-  });
-
-  it('keeps workspace renaming when no host home prefix was resolved', () => {
-    const cwd = '/workspace/repo';
-    expect(
-      redactText(`work in ${cwd}; leave /home/me/x alone`, cwd, []).value,
-    ).toBe('work in .; leave /home/me/x alone');
-  });
-
-  it('reports redacted:false when a path rule found nothing to rename', () => {
-    expect(redactText('nothing to rename here', '/workspace/repo', ['/home/me']).redacted)
-      .toBe(false);
-  });
-
   it('marks redacted:true only when a rule actually fired, false for ordinary text', () => {
     const { publisher, projection, agent } = harness();
     projectPrompt(projection, agent, 'nothing sensitive here at all');
@@ -503,10 +375,6 @@ describe('conversation projection: content visible after redaction is unchanged,
       error: null,
     });
 
-    // A newline serialized into JSON is the two characters `\n`, whose `n` is a
-    // word character: without the escape counting as a boundary, every line
-    // after the first would keep its secret. And a value must stop at that same
-    // escape, or one redacted line swallows the lines below it.
     const tool = activityOf(publisher);
     expect(tool.kind === 'tool.call' && tool.arguments_json).toBe(JSON.stringify({
       file_path: '.env',
@@ -515,7 +383,7 @@ describe('conversation projection: content visible after redaction is unchanged,
     expect(tool.kind === 'tool.call' && tool.redacted).toBe(true);
   });
 
-  it('keeps a redacted argument payload parsable as JSON', () => {
+  it('keeps a redacted argument payload parsable as JSON, by construction', () => {
     const { publisher, projection, agent } = harness();
     projection.projectActivity(agent, {
       kind: 'tool.call',
@@ -554,10 +422,9 @@ describe('conversation projection: content visible after redaction is unchanged,
       invocation: null,
       items: [],
       status: 'completed',
-      // A literal backslash serializes as `\\`. Read one character at a time it
-      // looks like the `\n` that ends a value, so the escaped pair is taken as
-      // a unit — otherwise this leaks `\name`, and a value ending in a
-      // backslash steals the quote that closes its string.
+      // Serialized, a literal backslash is `\\` and a value ending in one runs
+      // into the quote that closes its string. Neither is a problem the walk
+      // has: it hands over the leaf, which is the text the runtime wrote.
       arguments: { command: 'password=C:\\name', fallback: 'password=abc\\' },
       result: null,
       error: null,
